% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{add_coverage_heatmap}
\alias{add_coverage_heatmap}
\title{add_coverage_heatmap}
\usage{
add_coverage_heatmap(p, mat, x = iheatmapr:::default_x(mat),
  groups = p$row_groups, signal = rowMeans(mat, na.rm = TRUE),
  plot_signal = TRUE, name = "Coverage", signal_name = "Avg. Coverage",
  summary = TRUE, scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = 1, show_xlabels = TRUE, start = x[1],
  end = default_end(x), xlab = "Position", font = list(size = 8), ...)
}
\arguments{
\item{x}{x axis labels}

\item{groups}{pre-determined groups for rows}

\item{signal}{signal along row}

\item{plot_signal}{add an annotation heatmap showing the average signal? default is TRUE}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean, default is TRUE}

\item{scale_method}{how to scale matrix before displaying in heatmap}

\item{pct}{percentile to use if scale_method is "PercentileMax"}

\item{scale_factor}{scale_factor to use if scale_method is "scalar"}

\item{show_xlabels}{show xlabels?  default is TRUE}

\item{start}{label for start of x range}

\item{end}{label for end of x range}

\item{xlab}{x axis label}

\item{font}{list of font attributes}
}
\description{
add_coverage_heatmap
}
\author{
Alicia Schep
}

