% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{add_dist_to_tss}
\alias{add_dist_to_tss}
\title{add_dist_to_tss}
\usage{
add_dist_to_tss(p, ranges, tss, ...)
}
\arguments{
\item{p}{}

\item{ranges}{}

\item{tss}{}

\item{...}{}
}
\description{
add_dist_to_tss
}
