% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{adjust_y_domains}
\alias{adjust_y_domains}
\title{adjust_y_domains
Meant to adjust yaxes domains, primarily for mesocale views. Adjusts domains
to properly view annotations with each subplot}
\usage{
adjust_y_domains(plotly_obj, ax_info, heights, sort_decrease = TRUE)
}
\arguments{
\item{plotly_obj}{plotly: built plotly object to be modified}

\item{ax_info}{data.frame: contains relevant axes information. created using 
get_subplot_ax_info}

\item{heights}{numeric: heights of each track, including annotation tracks}
}
\value{
plotly_obj
}
\description{
adjust_y_domains
Meant to adjust yaxes domains, primarily for mesocale views. Adjusts domains
to properly view annotations with each subplot
}
\author{
Justin Finkle
}
