% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_plus_heatmap_shiny.R
\name{aggregate_shiny_plus}
\alias{aggregate_shiny_plus}
\title{aggregate_shiny_plus}
\usage{
aggregate_shiny_plus(mats, regions, track_function, region_widths = 50000,
  sample_groups = NULL, colors = "Dark2",
  aggregate_scale_method = c("none", "localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar"), showlegend = TRUE,
  x = iheatmapr:::default_x(mats[[1]]),
  y = iheatmapr:::default_y(mats[[1]]), include = 1000,
  include_method = c("signal", "first", "random"), row_order = c("none",
  "hclust", "kmeans", "groups", "signal"), k = NULL, row_groups = NULL,
  clust_dist = stats::dist, signal = lapply(mats, rowMeans, na.rm = TRUE),
  plot_signal = TRUE, name = "Coverage", signal_name = "Avg. Coverage",
  summary = TRUE, heatmap_scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factors = rep(1, length(mats)), ticktext = TRUE, start = x[1],
  end = default_end(x), xlab = "Position", font = list(size = 8),
  title = "Aggregate Plot Linked to Heatmap", options = list(height = 1400),
  ...)
}
\arguments{
\item{mats}{}

\item{regions}{}

\item{track_function}{}

\item{region_widths}{}

\item{sample_groups}{vector of groups to which matrices belong}

\item{colors}{either RColorbrewer pallete name or vector of colors}

\item{aggregate_scale_method}{}

\item{showlegend}{show the legend?}

\item{x}{}

\item{y}{}

\item{include}{}

\item{include_method}{}

\item{row_order}{}

\item{k}{}

\item{row_groups}{}

\item{clust_dist}{}

\item{signal}{}

\item{plot_signal}{}

\item{name}{}

\item{signal_name}{}

\item{summary}{}

\item{heatmap_scale_method}{}

\item{pct}{percentile argument if PercentileMax chosen as scale_method}

\item{scale_factors}{scale_factors if "scalar" chosen as scale_method}

\item{ticktext}{}

\item{start}{}

\item{end}{}

\item{xlab}{axis name for x axis}

\item{font}{}

\item{title}{}

\item{options}{}

\item{...}{}
}
\description{
aggregate_shiny_plus
}
\details{
More details will go here.
}
\author{
Alicia Schep
}
