% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{arrow_helper}
\alias{arrow_helper}
\title{arrow_helper
Draw lines that make arrows on the intron shapes}
\usage{
arrow_helper(arrow_start, strand, arrowlen, arrowheight, y, yref)
}
\arguments{
\item{arrow_start}{}

\item{strand}{}

\item{arrowlen}{}

\item{arrowheight}{}

\item{y}{}

\item{yref}{}
}
\value{

}
\description{
arrow_helper
Draw lines that make arrows on the intron shapes
}
