% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly.R
\name{browserly_cvg_track}
\alias{browserly_cvg_track}
\title{browserly_cvg_track
Make an interactive coverage track plot that can stand alone or be used as a subplot}
\usage{
browserly_cvg_track(cvg, track_name, type = c("scatter", "heatmap"),
  fill = "tozeroy", mode = "lines", showlegend = c(TRUE, FALSE),
  colors = NULL, ...)
}
\arguments{
\item{cvg}{GRanges: mcols should include the data to be plotted}

\item{track_name}{str: unique identifier for the plot. This is used to find appropriate axes when controlling annotations and domains}

\item{type}{str: type of plot}

\item{fill}{}

\item{mode}{}

\item{showlegend}{}

\item{colors}{}

\item{...}{additional arguments passed to plotly function "add_trace"}
}
\value{
plotly object
}
\description{
browserly_cvg_track
Make an interactive coverage track plot that can stand alone or be used as a subplot
}
