% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\docType{methods}
\name{coverage_heatmap}
\alias{coverage_heatmap}
\alias{add_coverage_heatmap}
\alias{coverage_heatmap,matrix-method}
\alias{coverage_heatmap-matrix}
\alias{add_coverage_heatmap,IheatmapHorizontal,matrix-method}
\alias{IheatmapHorizontal-matrix}
\alias{coverage_heatmap,list-method}
\alias{coverage_heatmap-list}
\alias{add_coverage_heatmap,IheatmapHorizontal,list-method}
\alias{IheatmapHorizontal-list}
\title{Coverage Heatmap}
\usage{
coverage_heatmap(data, ...)

add_coverage_heatmap(p, data, ...)

\S4method{coverage_heatmap}{matrix}(data, x = default_x(data),
  y = default_y(data), row_order = c("signal", "hclust", "kmeans", "groups",
  "none"), k = NULL, groups = NULL, clust_dist = stats::dist,
  signal = log10rowMeans(data), plot_signal = TRUE, name = "Coverage",
  signal_name = "Avg. (log10)", summary = TRUE,
  scale_method = c("localRms", "localMean", "localNonZeroMean",
  "PercentileMax", "scalar", "none"), pct = 0.95, scale_factor = 1,
  show_xlabels = TRUE, start = x[1], end = default_end(x),
  col_title = "Position", font = list(size = 10), ...)

\S4method{add_coverage_heatmap}{IheatmapHorizontal,matrix}(p, data,
  x = default_x(data), signal = log10rowMeans(data), plot_signal = TRUE,
  name = "Coverage", signal_name = "Avg. (log10)", summary = TRUE,
  scale_method = c("localRms", "localMean", "localNonZeroMean",
  "PercentileMax", "scalar", "none"), pct = 0.95, scale_factor = 1,
  show_xlabels = TRUE, start = x[1], end = default_end(x),
  col_title = "Position", font = list(size = 10), ...)

\S4method{coverage_heatmap}{list}(data, x = default_x(data[[1]]),
  y = default_y(data[[1]]), row_order = c("signal", "hclust", "kmeans",
  "groups", "none"), k = NULL, groups = NULL, clust_dist = stats::dist,
  cluster_by = c("first", "all"), signal = lapply(data, log10rowMeans),
  plot_signal = TRUE, name = "Coverage", signal_name = "Avg. (log10)",
  summary = TRUE, scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = 1, show_xlabels = TRUE, start = x[1],
  end = default_end(x), col_title = "Position", font = list(size = 10),
  ...)

\S4method{add_coverage_heatmap}{IheatmapHorizontal,list}(p, data,
  x = default_x(data[[1]]), signal = lapply(data, log10rowMeans),
  plot_signal = TRUE, name = "Coverage", signal_name = "Avg. (log10)",
  summary = TRUE, scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = 1, show_xlabels = TRUE, start = x[1],
  end = default_end(x), col_title = "Position", font = list(size = 10),
  ...)
}
\arguments{
\item{data}{single coverage matrix or list of coverage matrices}

\item{...}{}

\item{p}{}

\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{clust_dist}{distance function to use for clustering}

\item{signal}{signal along row}

\item{plot_signal}{add an annotation heatmap showing the average signal? 
default is TRUE}

\item{name}{name of colorbar}

\item{signal_name}{}

\item{summary}{make summary plot, boolean, default is TRUE}

\item{scale_method}{how to scale matrix before displaying in heatmap}

\item{pct}{percentile to use if scale_method is "PercentileMax"}

\item{scale_factor}{scale_factor to use if scale_method is "scalar"}

\item{show_xlabels}{show xlabels?  default is TRUE}

\item{start}{label for start of x range}

\item{end}{label for end of x range}

\item{col_title}{}

\item{font}{list of font attributes}

\item{cluster_by}{}
}
\value{
iheatmap object
}
\description{
Coverage Heatmap
}
\author{
Alicia Schep
}
