% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChipseq_functions.R
\name{expandSampleTable}
\alias{expandSampleTable}
\title{Expand sample table with paths to useful files}
\usage{
expandSampleTable(dat.samples, dir.bed)
}
\arguments{
\item{dat.samples}{Input sample table with minimal columns Sample.Name and Dir.ngs_pipeline}

\item{dir.bed}{}
}
\value{
data frame including additional columns
}
\description{
Expand sample table with paths to useful files
}
\details{
Expand sample table with paths to useful files
}
\author{
Suchit Jhunjhunwala
}
