% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{log10rowMeans}
\alias{log10rowMeans}
\title{log10rowMeans}
\usage{
log10rowMeans(x, na.rm = TRUE, pseudo = 1)
}
\arguments{
\item{x}{}

\item{na.rm}{}

\item{pseudo}{}
}
\description{
log10rowMeans
}
