% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{log_scale_range}
\alias{log_scale_range}
\title{Title}
\usage{
log_scale_range(center, chromosome, n = 10, extension = 1e+06,
  min_bin = 100)
}
\arguments{
\item{n}{}

\item{gr}{}
}
\description{
Title
}
