% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gviz_wrappers.R
\name{make_snp_track}
\alias{make_snp_track}
\title{Make SNP Track}
\usage{
make_snp_track(snp_gr, target_range, kwargs = list())
}
\arguments{
\item{snp_gr}{}

\item{target_range}{}

\item{kwargs}{}
}
\description{
Make SNP Track
}
\author{
Justin Finkle
}
