% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gviz_wrappers.R
\name{make_tss_track}
\alias{make_tss_track}
\title{Make TSS Track}
\usage{
make_tss_track(tss, target_range, chr, kwargs = list())
}
\arguments{
\item{tss}{}

\item{target_range}{}

\item{chr}{}
}
\description{
Make TSS Track
}
\author{
Justin Finkle
}
