% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{modify_y}
\alias{modify_y}
\title{modify_y
Modify y axes features such as titles, range, etc. This is a HIGHLY specific
function for making trackviews look nice.}
\usage{
modify_y(plotly_obj, ax_info, ann_ax, type, sync_y = TRUE,
  native_title = FALSE, title_rotation = 0)
}
\arguments{
\item{plotly_obj}{plotly: object to modify}

\item{ax_info}{data.frame: contains axes information so plot can be adjusted. 
See get_subplot_ax_info for more information}

\item{ann_ax}{character: the annotation axis of the form "yaxis[2-9]". 
These have different formatting requirements}

\item{type}{character: different types of plots are treated differently}

\item{sync_y}{logical: put all y-axes on same scale}

\item{native_title}{logical: use plotly titles. If FALSE, annotations are 
added to each subplot. Useful if titles are long to prevent title overlap.
 Margins are automatically adjusted}

\item{title_rotation}{numeric: angle of plot titles. Only used if 
native_title is FALSE}
}
\value{

}
\description{
modify_y
Modify y axes features such as titles, range, etc. This is a HIGHLY specific
function for making trackviews look nice.
}
\author{
Justin Finkle
}
