% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly.R
\name{plot_multiple_genes}
\alias{plot_multiple_genes}
\title{plot_multiple_genes
Plot several genes as separate tracks, each with multiple traces}
\usage{
plot_multiple_genes(genes, centered_cvg, centered_tx, hm_thresh = 4,
  stacking = "dense", type = "scatter", fill = "none", mode = "lines",
  sync_y = FALSE, xlabel = "TSS", gene_exprs = matrix(), ...)
}
\arguments{
\item{genes}{character: vector of gene names to be plotted}

\item{centered_cvg}{list: of GRanges coverage objects that are lined up at the TSS. See `get_centered_gene_info` for more detailes}

\item{centered_tx}{list: of GRanges annotation objects that are lined up at the TSS. See `get_centered_gene_info` for more detailes}

\item{hm_thresh}{integer: threshold above which coverage is plotted as a heatmap}

\item{stacking}{character: how to plot the annotations.}

\item{type}{character: the plot type. will override hm_thresh}

\item{fill}{character: fill for plotly traces}

\item{mode}{character: drawing mode for trace}

\item{sync_y}{logical: plots subplots on same y scale if TRUE}

\item{xlabel}{character: title for x axis}

\item{gene_exprs}{matrix: gene expression values to be plotted. rownames correspond to genes. columns are different samples and can have duplicate names. colnames are used for putting values in the correct boxplot}

\item{...}{}
}
\description{
plot_multiple_genes
Plot several genes as separate tracks, each with multiple traces
}
