% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{reflect_ranges}
\alias{reflect_ranges}
\title{reflect_ranges
Flip the range over a reference value}
\usage{
reflect_ranges(gr, reference = 0)
}
\arguments{
\item{gr}{GRanges: data to be flipped}

\item{reference}{integer: position around which values in the gr are reflected}
}
\value{

}
\description{
reflect_ranges
Flip the range over a reference value
}
\author{
Justin Finkle
}
