% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{set_tx_level}
\alias{set_tx_level}
\title{set_tx_level
Set the level that that transcripts will be displayed on}
\usage{
set_tx_level(tx_gr, stacking = c("dense", "squish"))
}
\arguments{
\item{tx_gr}{GRanges: contains transcript information}

\item{stacking}{str: dense collapses transcripts, squish expands them}
}
\value{

}
\description{
set_tx_level
Set the level that that transcripts will be displayed on
}

