% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChipseq_functions.R
\name{tssFromTxdb}
\alias{tssFromTxdb}
\title{get tss information from txdb}
\usage{
tssFromTxdb(txdb, regularize = TRUE)
}
\arguments{
\item{txdb}{input txdb object}

\item{regularize}{Logical, indicating whether only the regular chromosomes should be output}
}
\value{
tss granges
}
\description{
get tss from txdb
}
\details{
get tss information from txdb
}
\author{
Jinfeng Liu, Suchit Jhunjhunwala
}
